requirements.txt
fastapi
jinja2
uvicorn[standard]
sqlalchemy
alembic
python-multipart
python-jose[cryptography]
passlib
bcrypt==4.0.1
apscheduler